github_interactions/requirements.txt
GitPython==3.1.37
requests==2.31.0
PyGithub==2.1.1